Move executeShellCommand function to another file
